Move CoGBK tests into appropreate module.
